requirements_test.txt
coverage==4.4
dj_database_url>=0.4.1
django_extensions
pytest
pytest-django==4.2.0
pytest-mock
pytest-cov
